Moved test of computation of FD eigenvalues to test directory such that they are actually executed
